+++ genti-api/src/main/java/com/gt/genti/user/service/social/AppleOauthStrategy.java
@@ -2,54 +2,51 @@ package com.gt.genti.user.service.social;
 
 import static com.gt.genti.user.service.validator.UserValidator.*;
 
+import java.security.PublicKey;
+import java.util.Map;
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.gt.genti.error.ExpectedException;
+import com.gt.genti.error.ResponseCode;
 import com.gt.genti.jwt.JwtTokenProvider;
 import com.gt.genti.jwt.TokenGenerateCommand;
 import com.gt.genti.jwt.TokenResponse;
 import com.gt.genti.openfeign.apple.client.AppleApiClient;
-import com.gt.genti.openfeign.apple.service.AppleOAuthUserProvider;
+import com.gt.genti.openfeign.apple.dto.response.ApplePublicKeys;
+import com.gt.genti.openfeign.apple.service.AppleClaimsValidator;
+import com.gt.genti.openfeign.apple.service.AppleJwtParser;
 import com.gt.genti.openfeign.apple.service.AppleUserResponse;
+import com.gt.genti.openfeign.apple.service.PublicKeyGenerator;
 import com.gt.genti.user.dto.request.SocialLoginRequest;
 import com.gt.genti.user.dto.response.SocialLoginResponse;
 import com.gt.genti.user.model.OauthPlatform;
 import com.gt.genti.user.model.User;
 import com.gt.genti.user.repository.UserRepository;
-import com.gt.genti.user.service.UserSignUpService;
+import com.gt.genti.user.service.UserSignUpEventPublisher;
 import com.gt.genti.util.RandomUtil;
 
+import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
 public class AppleOauthStrategy implements SocialLoginStrategy {
 
-	@Value("${kakao.client-id}")
-	private String kakaoClientId;
-	@Value("${kakao.redirect-uri}")
-	private String kakaoRedirectUri;
-	@Value("${server.domain}")
-	private String serverBaseUri;
-	@Value("${server.port}")
-	private String serverPort;
-	@Value("${kakao.client-secret}")
-	private String kakaoClientSecret;
-
-	private final AppleApiClient appleApiClient;
-
-	private final AppleOAuthUserProvider appleOAuthUserProvider;
 	private final JwtTokenProvider jwtTokenProvider;
 	private final UserRepository userRepository;
-	private final UserSignUpService userSignUpService;
+	private final UserSignUpEventPublisher userSignUpEventPublisher;
+	private final AppleJwtParser appleJwtParser;
+	private final AppleApiClient appleApiClient;
+	private final PublicKeyGenerator publicKeyGenerator;
+	private final AppleClaimsValidator appleClaimsValidator;
 
 	@Override
 	@Transactional
 	public SocialLoginResponse login(SocialLoginRequest request) {
-		AppleUserResponse userResponse = appleOAuthUserProvider.getApplePlatformMember(request.getCode());
+		AppleUserResponse userResponse = getApplePlatformMember(request.getCode());
 		Optional<User> findUser = userRepository.findUserBySocialId(userResponse.getPlatformId());
 		User user;
 		boolean isNewUser = false;
@@ -63,7 +60,7 @@ public class AppleOauthStrategy implements SocialLoginStrategy {
 				.build());
 			user = newUser;
 			isNewUser = true;
-			userSignUpService.publishSignUpEvent(newUser);
+			userSignUpEventPublisher.publishSignUpEvent(newUser);
 		} else {
 			user = findUser.get();
 			user.resetDeleteAt();
@@ -83,4 +80,20 @@ public class AppleOauthStrategy implements SocialLoginStrategy {
 		return provider.equals(OauthPlatform.APPLE.getStringValue());
 	}
 
+	private AppleUserResponse getApplePlatformMember(String identityToken) {
+		Map<String, String> headers = appleJwtParser.parseHeaders(identityToken);
+		ApplePublicKeys applePublicKeys = appleApiClient.getApplePublicKeys();
+
+		PublicKey publicKey = publicKeyGenerator.generatePublicKey(headers, applePublicKeys);
+
+		Claims claims = appleJwtParser.parsePublicKeyAndGetClaims(identityToken, publicKey);
+		validateClaims(claims);
+		return new AppleUserResponse(claims.getSubject(), claims.get("email", String.class));
+	}
+
+	private void validateClaims(Claims claims) {
+		if (!appleClaimsValidator.isValid(claims)) {
+			throw ExpectedException.withLogging(ResponseCode.AppleOauthClaimInvalid);
+		}
+	}
 }
